requirements.txt
# requirements.txt
#
# Install these packages using:
# pip install -r requirements.txt

flask
flask-cors
google-generativeai
python-dotenv
